feat: inline conference sponsor form instead of external google forms
