+++ site/src/commonMain/libres/strings/conference_strings_en.xml
@@ -31,6 +31,7 @@
     <string name="conf_sponsor_caption">This event would not be possible without your help. Would you like to support the community and introduce your company to our entire audience?</string>
     <string name="conf_sponsor_im_interested">I'm interested</string>
     <string name="conf_sponsor_i_want_to_sponsor">I want to be an sponsor</string>
+    <string name="conf_sponsor_form_uid">327e2a365b</string>
 
     <!-- Conference Sponsorship -->
     <string name="conf_sponsorship_title">Interested in Sponsoring?</string>
