+++ site/src/jsMain/kotlin/web/android/dev/pe/pages/conf/components/sections/SponsorSection.kt
@@ -12,23 +12,18 @@ import org.jetbrains.compose.web.dom.H2
 import org.jetbrains.compose.web.dom.P
 import org.jetbrains.compose.web.dom.Text
 import strings.ResStrings
-import web.android.dev.pe.components.widgets.HeadingDecorator
+import web.android.dev.pe.components.utils.appendCurrentLanguage
 import web.android.dev.pe.components.widgets.OutlinePrimaryButtonVariant
 import web.android.dev.pe.components.widgets.PrimaryButton
 import web.android.dev.pe.pages.conf.components.layouts.ConferenceGridSection
 
 @Composable
-fun SponsorSection(
-    ctaTitle: String,
-    ctaPath: String,
-    sectionModifier: Modifier = Modifier,
-    modifier: Modifier = Modifier,
-) {
+fun SponsorSection(sectionModifier: Modifier = Modifier, modifier: Modifier = Modifier) {
     ConferenceGridSection(
         sectionModifier = sectionModifier,
         modifier = modifier,
         header = {
-            Details(ctaTitle, ctaPath)
+            Details()
         },
         content = {
             Sponsors()
@@ -38,15 +33,12 @@ fun SponsorSection(
 
 
 @Composable
-private fun Details(ctaTitle: String, ctaPath: String) {
-    H2 {
-        HeadingDecorator()
-        Text(ResStrings.conf_sponsor_title)
-    }
+private fun Details() {
+    H2 { Text(ResStrings.conf_sponsor_title) }
     P { Text(ResStrings.conf_sponsor_caption) }
     PrimaryButton(
-        text = ctaTitle,
-        href = ctaPath,
+        text = ResStrings.conf_sponsor_im_interested,
+        href = "/conf/sponsorship".appendCurrentLanguage(),
         modifier = Modifier.margin(topBottom = 16.px),
         variant = OutlinePrimaryButtonVariant,
     )
